Replace cron with node-scheduler
