--- crontab
@@ -1,2 +1,0 @@
-0 22 * * * set -a; . /var/app/saved_env; set +a; cd /var/app && /usr/local/bin/node bin/stalePullRequest.js >> /var/log/cron.log 2>&1
-# An empty line is required at the end of this file for a valid cron file.
