+++ lib/Settings.js
@@ -3,7 +3,6 @@
 var Cesium = require('cesium');
 var nconf = require('nconf');
 var Promise = require('bluebird');
-var fs = require('fs');
 
 var loadRepoConfig = require('./loadRepoConfig');
 var RepositorySettings = require('./RepositorySettings');
@@ -13,20 +12,6 @@ var defined = Cesium.defined;
 var Settings = {};
 
 var repositories = {};
-
-function dumpEnvironment(object, parentName) {
-    var envString = '';
-    Object.keys(object).forEach(function(name) {
-        var fullName = parentName === '' ? name : parentName + '__' + name;
-        if (typeof object[name] === 'object') {
-            var string = dumpEnvironment(object[name], fullName);
-            envString += string;
-        } else {
-            envString += fullName + ' = ' + object[name] + '\n';
-        }
-    });
-    return envString;
-}
 
 /** Check for necessary settings in `configPath` file and initialize `nconf`
  *
@@ -57,10 +42,6 @@ Settings.loadRepositoriesSettings = function (configPath) {
     if (repositoryNames.length === 0) {
         return Promise.reject('`repositories` must be non-empty');
     }
-
-    // Save the configuration to disk so that cron scripts can read from it
-    var environmentString = dumpEnvironment(nconf.get(), '');
-    fs.writeFile('saved_env', environmentString);
 
     return Promise.each(repositoryNames, function (name) {
         if (!/\//.test(name)) {
