+++ package.json
@@ -31,9 +31,10 @@
     "express": "^4.0.0",
     "handlebars": "^4.0.10",
     "nconf": "^0.8.4",
+    "node-schedule": "^1.3.0",
+    "parse-link-header": "^1.0.1",
     "request": "^2.81.0",
-    "request-promise": "^4.2.1",
-    "parse-link-header": "^1.0.1"
+    "request-promise": "^4.2.1"
   },
   "devDependencies": {
     "eslint": "^4.0.0",
